remove not used code
