+++ gemstone/src/swapper/cetus/clmm/error.rs
@@ -1,5 +1,6 @@
 use thiserror::Error;
 
+#[allow(unused)]
 #[derive(Debug, Clone, Copy, Error)]
 pub enum ErrorCode {
     #[error("Integer downcast overflow")]
