+++ gemstone/src/swapper/cetus/mod.rs
@@ -1,5 +1,4 @@
 mod client;
-#[allow(unused)]
 mod clmm;
 mod jsonrpc;
 mod models;
